feat: allow passing a configuration YAML
